try to migrate to Kotlin 1.5.0-RC

+++ server/src/test/kotlin/de/stefanbissell/starcruiser/ships/JumpHandlerTest.kt
@@ -177,7 +177,7 @@ class JumpHandlerTest {
     }
 
     private fun stepTime(seconds: Number) {
-        time.update(seconds.toDouble())
+        time.update(seconds)
         jumpHandler.update(time, physicsEngine, power)
     }
 }
