+++ server/src/test/kotlin/de/stefanbissell/starcruiser/ships/PlayerShipTest.kt
@@ -633,7 +633,7 @@ class PlayerShipTest {
     }
 
     private fun stepTime(seconds: Number) {
-        time.update(seconds.toDouble())
+        time.update(seconds)
         ship.update(time, physicsEngine, contactList)
     }
